Finalise right side layout and content
Add layout and formatting to the right side to make it achieve the
desired layout as depicted in the screenshot provided to us. Added
necessary content to make the website less boring and more interesting.

Make some quality of life changes to the html to make things easier to
both code and think through.